refactor: Refactor `login` service
